data-scraping-service/requirements.txt
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
dnspython==2.6.1
fonttools==4.54.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pymongo==4.9.1
pyparsing==3.1.4
python-bcb==0.3.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
schedule==1.2.2
six==1.16.0
sniffio==1.3.1
tzdata==2024.2
urllib3==2.2.3